Delete MLB VQA. Migrate VqaDict to save the vqa_v1 task.
