+++ parlai/agents/README.md
@@ -29,8 +29,3 @@ This directory contains a variety of different agents which use ParlAI's interfa
 - **seq2seq**: our RNN-based sequence to sequence model
 - **starspace**: embedding model
 - **transformer**: both generative and retrieval-based transformer models
-
-## Visual (+text) neural networks
-
-- **mlb_vqa**: visual question answering model
-- **vsepp_caption**: image captioning model
